v26.04.27: IM hell -> composition in Storage - documented Finder mode - size_t -> unsigned int in multiple places - updated input storage format, cov matrix is a single branch now with a stride - multiple variables, inits and branches renamed - removed functions that store dummy rows

+++ include/App/Parser.hxx
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <array>
-#include <cstdlib>
 #include <string>
 #include <vector>
 
@@ -51,7 +50,7 @@ class Parser {
         // -- input path
         settings.PathInputFiles = InputFiles;
         // -- n events limit
-        settings.LimitToNEvents = CLI_APP.get_option("-n")->as<size_t>();
+        settings.LimitToNEvents = CLI_APP.get_option("-n")->as<unsigned int>();
         // -- output path
         settings.PathOutputFile = CLI_APP.get_option("-o")->as<std::string>();
         if (settings.PathOutputFile.empty()) {
